requirements.txt
python-dotenv~=1.0.1
ipython~=8.18.1
pip~=25.0
wheel~=0.41.2
exceptiongroup~=1.2.2
pygments~=2.19.1
setuptools~=70.0.0
langgraph
langchain-core
langchain-openai
tavily-python
langgraph-checkpoint-sqlite